.Net Agents - Support single-agent / no-chat invocation of `OpenAIAssistantAgent` (#6946)
### Motivation and Context
<!-- Thank you for your contribution to the semantic-kernel repo!
Please help reviewers and future users, providing the following
information:
  1. Why is this change required?
  2. What problem does it solve?
  3. What scenario does it contribute to?
  4. If it fixes an open issue, please link to the issue here.
-->

Support pattern for direct invocation of `OpenAIAssistantAgent` in
alignment with `ChatCompletionAgent`

### Description
<!-- Describe your changes, the overall approach, the underlying design.
These notes will help understanding how your code works. Thanks! -->

- Add instance method `OpenAIAssistantAgent.InvokeAsync` for a specified
_thread-id_
- Add instance method `OpenAIAssistantAgent.AddChatMessageAsync` for a
specified _thread-id_
- Add thread management methods to: Create and Delete a thread.
- Refactored thread related actions to static class
`AssistantThreadActions` so it may be utilized by both
`OpenAIAssistantAgent` and `OpenAIAssistantChannel`
- Updated samples to demonstrate no-chat invocation pattern

### Contribution Checklist
<!-- Before submitting this PR, please make sure: -->

- [X] The code builds clean without any errors or warnings
- [X] The PR follows the [SK Contribution
Guidelines](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md)
and the [pre-submission formatting
script](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md#development-scripts)
raises no violations
- [X] All unit tests pass, and I have added new tests where possible
- [X] I didn't break anyone :smile:
